Move util.HashString to internal/hashcode.String
